Restructure to module
