+++ src/solver/mod.rs
@@ -1,0 +1,2 @@
+pub mod divide_and_conquer;
+pub mod optimal;
